requirements.txt
amqp==2.1.4
aniso8601==1.2.1
appdirs==1.4.3
billiard==3.5.0.2
celery==4.0.2
certifi==2017.4.17
chardet==3.0.4
click==6.7
Flask==0.12.2
Flask-RESTful==0.3.6
Flask-Script==2.0.5
idna==2.5
itsdangerous==0.24
Jinja2==2.9.6
kombu==4.0.2
MarkupSafe==1.0
packaging==16.8
pyparsing==2.2.0
python-dateutil==2.6.0
pytz==2017.2
PyYAML==3.12
redis==2.10.5
requests==2.17.3
simplejson==3.10.0
six==1.10.0
urllib3==1.21.1
vine==1.1.3
Werkzeug==0.12.2
